create wasm_client_tests.rs
